% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{create_button_html}
\alias{create_button_html}
\title{Create a button link from meta data}
\usage{
create_button_html(metadata)
}
\arguments{
\item{metadata}{The output of get_metadata}
}
\value{
HTML text defining a button with the title, url, and image from the website's meta data.
}
\description{
Create a button link from meta data
}
\examples{
get_metadata("https://www.eurekalert.org/pub_releases/2020-10/wfbm-srr100920.php") \%>\%
create_button_html()

}
