% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_metadata}
\alias{get_metadata}
\title{Scrape Open Graph Metadata from Websites}
\usage{
get_metadata(url)
}
\arguments{
\item{url}{Character string of a website url}
}
\value{
A tidy dataframe with the Open Graph metadata paramaters, such as
title, description, and image.
}
\description{
Scrape Open Graph Metadata from Websites
}
\examples{
get_metadata("https://www.eurekalert.org/pub_releases/2020-10/wfbm-srr100920.php")
}
